bases: remove support for core18
