--- snapcraft/plugins/v1/__init__.py
@@ -1,126 +1,0 @@
-# -*- Mode:Python; indent-tabs-mode:nil; tab-width:4 -*-
-#
-# Copyright 2022 Canonical Ltd.
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License version 3 as
-# published by the Free Software Foundation.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-"""Legacy support for v1 plugins."""
-
-
-from snapcraft_legacy.plugins.v1 import (
-    AutotoolsPlugin,
-    CatkinPlugin,
-    CatkinToolsPlugin,
-    CMakePlugin,
-    ColconPlugin,
-    CondaPlugin,
-    CrystalPlugin,
-    DotNetPlugin,
-    DumpPlugin,
-    FlutterPlugin,
-    GodepsPlugin,
-    GoPlugin,
-    GradlePlugin,
-    KBuildPlugin,
-    KernelPlugin,
-    MakePlugin,
-    MavenPlugin,
-    MesonPlugin,
-    NilPlugin,
-    NodePlugin,
-    PlainboxProviderPlugin,
-    PluginV1,
-    PythonPlugin,
-    QmakePlugin,
-    RubyPlugin,
-    RustPlugin,
-    autotools,
-    catkin,
-    catkin_tools,
-    cmake,
-    colcon,
-    conda,
-    crystal,
-    dotnet,
-    dump,
-    flutter,
-    go,
-    godeps,
-    gradle,
-    kbuild,
-    kernel,
-    make,
-    maven,
-    meson,
-    nil,
-    nodejs,
-    plainbox_provider,
-    python,
-    qmake,
-    ruby,
-    rust,
-)
-
-__all__ = [
-    "PluginV1",
-    "autotools",
-    "catkin",
-    "catkin_tools",
-    "cmake",
-    "colcon",
-    "conda",
-    "crystal",
-    "dotnet",
-    "dump",
-    "flutter",
-    "godeps",
-    "go",
-    "gradle",
-    "kbuild",
-    "kernel",
-    "make",
-    "maven",
-    "meson",
-    "nil",
-    "nodejs",
-    "plainbox_provider",
-    "python",
-    "qmake",
-    "ruby",
-    "rust",
-    "AutotoolsPlugin",
-    "CatkinPlugin",
-    "CatkinToolsPlugin",
-    "CMakePlugin",
-    "ColconPlugin",
-    "CondaPlugin",
-    "CrystalPlugin",
-    "DotNetPlugin",
-    "DumpPlugin",
-    "FlutterPlugin",
-    "GodepsPlugin",
-    "GoPlugin",
-    "GradlePlugin",
-    "KBuildPlugin",
-    "KernelPlugin",
-    "MakePlugin",
-    "MavenPlugin",
-    "MesonPlugin",
-    "NilPlugin",
-    "NodePlugin",
-    "PlainboxProviderPlugin",
-    "PythonPlugin",
-    "QmakePlugin",
-    "RustPlugin",
-    "RubyPlugin",
-]
